fix: 404/400を正しく返すよう例外ハンドラを補強し、テストをStudentControllerTestに集約
